sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flamepy"
version = "0.5.0"
description = "Python SDK for Flame, a distributed system for Agentic AI"
readme = "README.md"
requires-python = ">=3.12"
license-files = ["LICENSE"]
authors = [
  {name = "XFLOPS Authors", email = "support@xflops.io"}
]
keywords = ["distributed system", "agentic ai", "flame", "sdk"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  "grpcio>=1.50.0",
  "grpcio-tools>=1.50.0",
  "protobuf>=4.21.0",
  "pydantic>=2.11.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.35.0",
  "pyyaml>=6.0.0",
  "bson>=0.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.21.0",
  "black>=22.0.0",
  "isort>=5.0.0",
  "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = [
    "src"
]

[project.urls]
Homepage = "https://xflops.io"
Repository = "https://github.com/xflops/flame"
Issues = "https://github.com/xflops/flame/issues"

[tool.black]
line-length = 1024
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["flame"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "grpc.*",
    "grpc.aio.*",
    "grpc_tools.*",
]
ignore_missing_imports = true 

[dependency-groups]
dev = [
    "black>=25.1.0",
]

